return vmError and txError apart
