+++ fvm/bootstrap.go
@@ -114,10 +114,7 @@ func (b *BootstrapProcedure) Run(vm *VirtualMachine, ctx Context, sth *state.Sta
 
 	// initialize the account addressing state
 	b.accounts = state.NewAccounts(b.sth)
-	addressGenerator, err := state.NewStateBoundAddressGenerator(b.sth, ctx.Chain)
-	if err != nil {
-		panic(fmt.Sprintf("failed to create address generator: %s", err.Error()))
-	}
+	addressGenerator := state.NewStateBoundAddressGenerator(b.sth, ctx.Chain)
 	b.addressGenerator = addressGenerator
 
 	service := b.createServiceAccount(b.serviceAccountPublicKey)
@@ -169,16 +166,18 @@ func (b *BootstrapProcedure) createServiceAccount(accountKey flow.AccountPublicK
 func (b *BootstrapProcedure) deployFungibleToken() flow.Address {
 	fungibleToken := b.createAccount()
 
-	err := b.vm.invokeMetaTransaction(
+	txError, vmErr := b.vm.invokeMetaTransaction(
 		b.ctx,
 		deployContractTransaction(fungibleToken, contracts.FungibleToken(), "FungibleToken"),
 		b.sth,
 		b.programs,
 	)
-	if err != nil {
-		panic(fmt.Sprintf("failed to deploy fungible token contract: %s", err.Error()))
+	if txError != nil {
+		panic(fmt.Sprintf("failed to deploy fungible token contract: %s", txError.Error()))
 	}
-
+	if vmErr != nil {
+		panic(fmt.Sprintf("failed to deploy fungible token contract: %s", vmErr.Error()))
+	}
 	return fungibleToken
 }
 
@@ -187,16 +186,18 @@ func (b *BootstrapProcedure) deployFlowToken(service, fungibleToken flow.Address
 
 	contract := contracts.FlowToken(fungibleToken.HexWithPrefix())
 
-	err := b.vm.invokeMetaTransaction(
+	txErr, vmErr := b.vm.invokeMetaTransaction(
 		b.ctx,
 		deployFlowTokenTransaction(flowToken, service, contract),
 		b.sth,
 		b.programs,
 	)
-	if err != nil {
-		panic(fmt.Sprintf("failed to deploy Flow token contract: %s", err.Error()))
+	if txErr != nil {
+		panic(fmt.Sprintf("failed to deploy Flow token contract: %s", txErr.Error()))
 	}
-
+	if vmErr != nil {
+		panic(fmt.Sprintf("failed to deploy Flow token contract: %s", vmErr.Error()))
+	}
 	return flowToken
 }
 
@@ -208,16 +209,18 @@ func (b *BootstrapProcedure) deployFlowFees(service, fungibleToken, flowToken fl
 		flowToken.HexWithPrefix(),
 	)
 
-	err := b.vm.invokeMetaTransaction(
+	txErr, vmErr := b.vm.invokeMetaTransaction(
 		b.ctx,
 		deployFlowFeesTransaction(flowFees, service, contract),
 		b.sth,
 		b.programs,
 	)
-	if err != nil {
-		panic(fmt.Sprintf("failed to deploy fees contract: %s", err.Error()))
+	if txErr != nil {
+		panic(fmt.Sprintf("failed to deploy fees contract: %s", txErr.Error()))
 	}
-
+	if vmErr != nil {
+		panic(fmt.Sprintf("failed to deploy fees contract: %s", vmErr.Error()))
+	}
 	return flowFees
 }
 
@@ -228,15 +231,19 @@ func (b *BootstrapProcedure) deployStorageFees(service, fungibleToken, flowToken
 	)
 
 	// deploy storage fees contract on the service account
-	err := b.vm.invokeMetaTransaction(
+	txErr, vmErr := b.vm.invokeMetaTransaction(
 		b.ctx,
 		deployStorageFeesTransaction(service, contract),
 		b.sth,
 		b.programs,
 	)
-	if err != nil {
-		panic(fmt.Sprintf("failed to deploy storage fees contract: %s", err.Error()))
+	if txErr != nil {
+		panic(fmt.Sprintf("failed to deploy storage fees contract: %s", txErr.Error()))
 	}
+	if vmErr != nil {
+		panic(fmt.Sprintf("failed to deploy storage fees contract: %s", vmErr.Error()))
+	}
+
 }
 
 func (b *BootstrapProcedure) deployServiceAccount(service, fungibleToken, flowToken, feeContract flow.Address) {
@@ -247,14 +254,17 @@ func (b *BootstrapProcedure) deployServiceAccount(service, fungibleToken, flowTo
 		service.HexWithPrefix(),
 	)
 
-	err := b.vm.invokeMetaTransaction(
+	txErr, vmErr := b.vm.invokeMetaTransaction(
 		b.ctx,
 		deployContractTransaction(service, contract, "FlowServiceAccount"),
 		b.sth,
 		b.programs,
 	)
-	if err != nil {
-		panic(fmt.Sprintf("failed to deploy service account contract: %s", err.Error()))
+	if txErr != nil {
+		panic(fmt.Sprintf("failed to deploy service account contract: %s", txErr.Error()))
+	}
+	if vmErr != nil {
+		panic(fmt.Sprintf("failed to deploy service account contract: %s", vmErr.Error()))
 	}
 }
 
@@ -262,14 +272,17 @@ func (b *BootstrapProcedure) mintInitialTokens(
 	service, fungibleToken, flowToken flow.Address,
 	initialSupply cadence.UFix64,
 ) {
-	err := b.vm.invokeMetaTransaction(
+	txErr, vmErr := b.vm.invokeMetaTransaction(
 		b.ctx,
 		mintFlowTokenTransaction(fungibleToken, flowToken, service, initialSupply),
 		b.sth,
 		b.programs,
 	)
-	if err != nil {
-		panic(fmt.Sprintf("failed to mint initial token supply: %s", err.Error()))
+	if txErr != nil {
+		panic(fmt.Sprintf("failed to mint initial token supply: %s", txErr.Error()))
+	}
+	if vmErr != nil {
+		panic(fmt.Sprintf("failed to mint initial token supply: %s", vmErr.Error()))
 	}
 }
 
@@ -279,28 +292,34 @@ func (b *BootstrapProcedure) setupFees(
 	addressCreationFee,
 	minimumStorageReservation cadence.UFix64,
 ) {
-	err := b.vm.invokeMetaTransaction(
+	txErr, vmErr := b.vm.invokeMetaTransaction(
 		b.ctx,
 		setupFeesTransaction(service, transactionFee, addressCreationFee, minimumStorageReservation),
 		b.sth,
 		b.programs,
 	)
-	if err != nil {
-		panic(fmt.Sprintf("failed to setup fees: %s", err.Error()))
+	if txErr != nil {
+		panic(fmt.Sprintf("failed to setup fees: %s", txErr.Error()))
+	}
+	if vmErr != nil {
+		panic(fmt.Sprintf("failed to setup fees: %s", vmErr.Error()))
 	}
 }
 
 func (b *BootstrapProcedure) setupStorageForServiceAccounts(
 	service, fungibleToken, flowToken, feeContract flow.Address,
 ) {
-	err := b.vm.invokeMetaTransaction(
+	txErr, vmErr := b.vm.invokeMetaTransaction(
 		b.ctx,
 		setupStorageForServiceAccountsTransaction(service, fungibleToken, flowToken, feeContract),
 		b.sth,
 		b.programs,
 	)
-	if err != nil {
-		panic(fmt.Sprintf("failed to setup storage for service accounts: %s", err.Error()))
+	if txErr != nil {
+		panic(fmt.Sprintf("failed to setup storage for service accounts: %s", txErr.Error()))
+	}
+	if vmErr != nil {
+		panic(fmt.Sprintf("failed to setup storage for service accounts: %s", vmErr.Error()))
 	}
 }
