+++ fvm/fvm.go
@@ -82,16 +82,8 @@ func (vm *VirtualMachine) GetAccount(ctx Context, address flow.Address, v state.
 //
 // Errors that occur in a meta transaction are propagated as a single error that can be
 // captured by the Cadence runtime and eventually disambiguated by the parent context.
-func (vm *VirtualMachine) invokeMetaTransaction(ctx Context, tx *TransactionProcedure, sth *state.StateHolder, programs *programs.Programs) error {
+func (vm *VirtualMachine) invokeMetaTransaction(ctx Context, tx *TransactionProcedure, sth *state.StateHolder, programs *programs.Programs) (txError error, vmError error) {
 	invocator := NewTransactionInvocator(zerolog.Nop())
-	err := invocator.Process(vm, &ctx, tx, sth, programs)
-	if err != nil {
-		return err
-	}
-
-	if tx.Err != nil {
-		return tx.Err
-	}
-
-	return nil
+	txError, vmError = invocator.Process(vm, &ctx, tx, sth, programs)
+	return
 }
